# openapi.yaml
openapi: 3.0.3
info:
  title: Byzantium API
  version: 0.1.0
  description: >
    Byzantium is an autonomous-agent trust and compliance gateway. It evaluates
    whether an agent is authorised to perform a given action by enforcing
    operator-defined spend mandates, checking behavioural reputation scores, and
    issuing short-lived ML-DSA–signed PassTokens that payment rails can verify
    in microseconds without calling back to Byzantium. The API also provides
    mandate lifecycle management, liability-receipt recording with Merkle batch
    anchoring, DID-based agent identity, regulator-grade audit export, EIP-3009
    and Solana SPL payment-rail verification, and operator API-key management.
servers:
  - url: http://localhost:8080
    description: Local development server (no TLS)
  - url: https://api.byzantium.example.com
    description: Production — TLS terminated by nginx; same path structure
security:
  - BearerAuth: []
...
paths:
...
  /v1/trust-check:
    post:
      tags: [Trust]
      operationId: trustCheck
      summary: Evaluate agent trust for an action
      description: >
        Runs the full trust pipeline — mandate compliance check, reputation
        threshold evaluation, and (on pass) issues a short-lived ML-DSA–signed
        PassToken that the calling rail can verify offline. Target latency is
        under 200 ms.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrustCheckRequest'
            example:
              agent_did: "did:byz:agent:550e8400-e29b-41d4-a716-446655440000"
              action_type: payment
              amount_cents: 5000
              counterparty:
                id: "merchant-42"
                chain: "base"
                address: "0xAbCd1234..."
              rail_id: "eip3009-base"
              idempotency_key: "req-abc-123"
      responses:
        '200':
          description: Trust evaluation completed (verdict may be PASS, FLAG, or BLOCK)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrustCheckResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'

  # ── Mandates ─────────────────────────────────────────────────────────────────
  /v1/mandates:
    post:
      tags: [Mandates]
      operationId: createMandate
      summary: Create a spend mandate for an agent
      description: >
        Creates a new cryptographically-signed SpendMandate defining what an
        agent is allowed to do; the mandate root is the SHA-256 Merkle root
        over all fields and is used for ZK range/membership proofs.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMandateRequest'
            example:
              agent_did: "did:byz:agent:550e8400-e29b-41d4-a716-446655440000"
              operator_id: "op-acme"
              counterparty_whitelist: ["merchant-42", "vendor-7"]
              allowed_action_types: ["payment", "api_call"]
              per_tx_cap_cents: 10000
              daily_cap_cents: 500000
              valid_days: 30
      responses:
        '200':
          description: Mandate created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateMandateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/mandates/{id}:
    get:
      tags: [Mandates]
      operationId: getMandate
      summary: Retrieve a mandate by ID
      description: >
        Returns the full SpendMandate record including its Merkle root and
        validity window; prefers the persistent store, falls back to in-memory.
      parameters:
        - $ref: '#/components/parameters/MandateId'
      responses:
        '200':
          description: Mandate found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpendMandate'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/mandates/{id}/revoke:
    post:
      tags: [Mandates]
      operationId: revokeMandate
      summary: Revoke a mandate immediately
      description: >
        Marks the mandate as revoked in both the persistent store and the
        in-memory engine; any subsequent trust-checks for this agent will fail
        compliance.
      parameters:
        - $ref: '#/components/parameters/MandateId'
      responses:
        '200':
          description: Mandate revoked
          content:
            application/json:
              schema:
                type: object
                required: [revoked]
                properties:
                  revoked:
                    type: string
                    format: uuid
                    description: UUID of the revoked mandate
              example:
                revoked: "550e8400-e29b-41d4-a716-446655440000"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ── Receipts ─────────────────────────────────────────────────────────────────
  /v1/receipts:
    post:
      tags: [Receipts]
      operationId: createReceipt
      summary: Record a signed liability receipt
      description: >
        Creates an ML-DSA–signed LiabilityReceipt for a completed agent action,
        persists it to PostgreSQL, and adds it to the rolling Merkle batch;
        if the batch reaches its size threshold it is automatically sealed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReceiptRequest'
            example:
              agent_did: "did:byz:agent:550e8400-e29b-41d4-a716-446655440000"
              action_type: payment
              counterparty:
                id: "merchant-42"
                chain: "base"
                address: "0xAbCd1234..."
              amount_cents: 5000
              outcome: success
              mandate_id: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
              rail_id: "eip3009-base"
      responses:
        '200':
          description: Receipt created and signed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LiabilityReceipt'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/receipts/{id}/proof:
    get:
      tags: [Receipts]
      operationId: inclusionProof
      summary: Get Merkle inclusion proof for a receipt
      description: >
        Returns the batch ID and Merkle root for the batch that contains the
        given receipt; full Merkle path generation requires rebuilding the batch
        tree from stored receipts.
      parameters:
        - $ref: '#/components/parameters/ReceiptId'
      responses:
        '200':
          description: Proof info for the receipt's batch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InclusionProofResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/batches/{id}/seal:
    post:
      tags: [Receipts]
      operationId: sealBatch
      summary: Force-seal the current receipt batch
      description: >
        Immediately seals the in-progress Merkle batch regardless of size,
        persists it to PostgreSQL, and assigns all pending receipts to this
        batch; useful for end-of-period anchoring.
      parameters:
        - $ref: '#/components/parameters/BatchId'
      responses:
        '200':
          description: Batch sealed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SealBatchResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'

  # ── Audit ─────────────────────────────────────────────────────────────────────
  /v1/audit/receipts:
    get:
      tags: [Audit]
      operationId: listAuditReceipts
      summary: Export receipts with Merkle proofs for audit
      description: >
        Returns paginated receipt records with their Merkle inclusion proofs so
        regulators and insurers can independently verify each receipt without
        trusting Byzantium; requires an audit-scoped Bearer key.
      parameters:
        - name: agent_did
          in: query
          required: false
          schema:
            type: string
          description: Filter receipts to a specific agent DID
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: >
            ISO 8601 UTC lower bound (inclusive) on receipt timestamp
            (e.g. `2024-01-01T00:00:00Z`)
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: ISO 8601 UTC upper bound (inclusive) on receipt timestamp
        - name: batch_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter to receipts belonging to a specific batch (in-memory mode only)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
          description: Maximum number of records to return (capped at 1000)
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: >
            Opaque keyset-pagination cursor from the previous page's
            `next_cursor` field (base64-encoded UUID string); omit to start
            from the beginning
      responses:
        '200':
          description: Paginated audit receipts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditReceiptsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/audit/batches/{id}:
    get:
      tags: [Audit]
      operationId: getBatchProof
      summary: Retrieve Merkle root for a sealed batch
      description: >
        Returns the Merkle root, receipt count, and seal timestamp for a given
        batch, enabling offline verification of batch integrity.
      parameters:
        - $ref: '#/components/parameters/BatchId'
      responses:
        '200':
          description: Batch proof details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchProofResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ── Identity ──────────────────────────────────────────────────────────────────
  /v1/agents:
    post:
      tags: [Identity]
      operationId: registerAgent
      summary: Register a new agent DID
      description: >
        Generates a fresh Dilithium (ML-DSA) keypair and a deterministic DID for
        the agent, stores the DID document, and returns the DID and public key
        hex so the operator can embed it in the agent's TEE configuration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterAgentRequest'
            example:
              operator_id: "op-acme"
              kyb_proof: "eyJhbGciOiJFZERTQSJ9..."
      responses:
        '200':
          description: Agent registered; DID and public key returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterAgentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/agents/{did}:
    get:
      tags: [Identity]
      operationId: getAgent
      summary: Resolve an agent DID document
      description: >
        Returns the DID document for the given agent including operator, public
        key, KYB status, and active flag; prefers the persistent store over the
        in-memory resolver.
      parameters:
        - $ref: '#/components/parameters/AgentDid'
      responses:
        '200':
          description: Agent DID document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentDocument'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/agents/{did}/deactivate:
    post:
      tags: [Identity]
      operationId: deactivateAgent
      summary: Deactivate an agent permanently
      description: >
        Marks the agent's DID document as inactive in both the persistent store
        and the in-memory resolver; subsequent trust-checks for this DID will be
        blocked.
      parameters:
        - $ref: '#/components/parameters/AgentDid'
      responses:
        '200':
          description: Agent deactivated
          content:
            application/json:
              schema:
                type: object
                required: [deactivated]
                properties:
                  deactivated:
                    type: string
                    description: The DID that was deactivated
              example:
                deactivated: "did:byz:agent:550e8400-e29b-41d4-a716-446655440000"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ── Payments ──────────────────────────────────────────────────────────────────
  /v1/payments/eip3009/verify:
    post:
      tags: [Payments]
      operationId: verifyEip3009
      summary: Verify an EIP-3009 transfer authorization
      description: >
        Validates an EIP-3009 TransferWithAuthorization: checks signature
        recovery, time window, nonce consumption on-chain, and runs an internal
        Byzantium trust-check on the sender before the agent action proceeds.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Eip3009VerifyRequest'
            example:
              authorization:
                token: "0x833589fcd6edb6e08f4c7c32d4f71b54bda02913"
                from: "0xSenderAddress"
                to: "0xRecipientAddress"
                value: 1000000
                valid_after: 1700000000
                valid_before: 1700086400
                nonce: "0x0000000000000000000000000000000000000000000000000000000000000001"
                signature: "0xabc123...65bytes"
              min_trust_score: 700
      responses:
        '200':
          description: Authorization verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Eip3009VerifyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/payments/solana/verify:
    post:
      tags: [Payments]
      operationId: verifySolana
      summary: Verify a Solana SPL token transfer
      description: >
        Confirms that a finalized Solana transaction matches the expected sender,
        recipient, mint, and amount, then runs an internal Byzantium trust-check
        on the sender wallet.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SolanaTransferProof'
            example:
              signature: "5KtPn1LGuxhFCPFd1...base58txsig"
              from_wallet: "9WzDXwBbmkg8ZTbNMqUxvQRAyrZzDsGYdLVL9zYtAWWM"
              to_wallet: "DRpbCBMxVnDK7maPM5tGv6MvB3v1sRMC86PZ8okm21hy"
              mint: "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v"
              amount: 1000000
              cluster: "mainnet-beta"
      responses:
        '200':
          description: Solana transfer verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolanaVerifyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ── Keys ──────────────────────────────────────────────────────────────────────
  /v1/keys:
    post:
      tags: [Keys]
      operationId: createKey
      summary: Create a new operator API key
      description: >
        Creates a new API key scoped to the authenticated operator; the raw key
        value is returned exactly once in this response and is never retrievable
        again — store it securely.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateKeyRequest'
            example:
              label: "production-agent-runner"
              scopes: ["trust:check", "receipts:write"]
              expires_at: "2026-12-31T23:59:59Z"
      responses:
        '200':
          description: Key created; raw value shown once
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateKeyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    get:
      tags: [Keys]
      operationId: listKeys
      summary: List API keys for the operator
      description: >
        Returns all API keys (active and revoked) for the authenticated operator;
        the raw key value is never included.
      responses:
        '200':
          description: List of API key records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKeyInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/keys/{id}:
    delete:
      tags: [Keys]
      operationId: revokeKey
      summary: Revoke an API key by ID
      description: >
        Permanently revokes the specified API key for the authenticated operator;
        any in-flight requests using the key will be rejected after the next
        auth middleware cycle.
      parameters:
        - $ref: '#/components/parameters/KeyId'
      responses:
        '200':
          description: Key revoked
          content:
            application/json:
              schema:
                type: object
                required: [status, id]
                properties:
                  status:
                    type: string
                    enum: [revoked]
                  id:
                    type: string
                    format: uuid
              example:
                status: "revoked"
                id: "550e8400-e29b-41d4-a716-446655440000"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  # ── Observability ─────────────────────────────────────────────────────────────

  # ── Underwriting ───────────────────────────────────────────────────────────
  /v1/principals:
    post:
      tags: [Underwriting]
      summary: Bind an agent to a KYC-verified principal
      description: >
        Nothing is underwritten before this. Standing is a gate rather than a
        score: a verified principal raises the ceiling an agent may eventually
        reach, but never earns the limit on its own. Limits also consolidate at
        the principal, so registering more agents divides a ceiling instead of
        multiplying it.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agent_did, principal_ref, kyc_tier, sanctions_clear]
              properties:
                agent_did: { type: string, example: "did:byz:trader-01" }
                principal_ref:
                  type: string
                  description: Pseudonymous reference to the principal, never raw identity.
                  example: "sha256:acme-treasury"
                kyc_tier:
                  type: string
                  enum: [unverified, basic, verified, institutional]
                sanctions_clear: { type: boolean }
                jurisdiction: { type: string, example: "SG" }
                entity_age_days: { type: integer, example: 900 }
      responses:
        "200":
          description: Bound
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent_did: { type: string }
                  principal_ref: { type: string }
                  kyc_tier: { type: string }
                  eligible: { type: boolean }
                  agents_under_principal: { type: integer }
        "400": { description: principal_ref missing }
  /v1/limits/issue:
    post:
      tags: [Underwriting]
      summary: Underwrite the agent and issue a portable limit
      description: >
        Runs the five controls — standing gate, sublinear growth in settled
        value, experience cap, per-window rate cap, principal consolidation —
        and returns a signed attestation. `reasons` lists every control that
        shaped the number, on an approval as well as a refusal.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agent_did]
              properties:
                agent_did: { type: string }
                ccy:
                  type: string
                  enum: [USD, SGD, EUR, GBP, JPY]
                  default: USD
                chains:
                  type: array
                  items: { type: string }
                  description: Chains this limit may be presented on. Empty means any.
                  example: ["base", "solana"]
                asset_classes:
                  type: array
                  items: { type: string, enum: [stablecoin, major_fiat, volatile] }
                counterparty_classes:
                  type: array
                  items: { type: string }
                action_types:
                  type: array
                  items: { type: string }
      responses:
        "200":
          description: Decision. `issued` is false on a refusal, with `refusal` explaining why.
          content:
            application/json:
              schema:
                type: object
                properties:
                  issued: { type: boolean }
                  attestation:
                    $ref: "#/components/schemas/LimitAttestation"
                  tier: { type: string, example: "B2" }
                  score: { type: integer, example: 481 }
                  reasons:
                    type: array
                    items: { type: string }
                  refusal: { type: string }
        "400": { description: No registered principal for this agent }
...
  /v1/limits/settle:
    post:
      tags: [Underwriting]
      summary: Resolve a committed draw
      description: >
        Settling consumes window capacity; reporting a failure releases the
        reservation instead. Either outcome reaches the scorer, which is what
        closes the feedback loop between behaviour and future limits.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agent_did, amount_minor, currency, settled]
              properties:
                agent_did: { type: string }
                amount_minor: { type: integer }
                currency: { type: string, enum: [USD, SGD, EUR, GBP, JPY] }
                settled:
                  type: boolean
                  description: False releases the reservation rather than consuming the window.
                idempotency_key: { type: string }
      responses:
        "200":
          description: Current exposure after the update
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent_did: { type: string }
                  settled: { type: boolean }
                  at_risk_minor: { type: integer }
                  window_used_minor: { type: integer }
...
  /v1/runtimes/revoke:
    post:
      tags: [Provenance]
      summary: Revoke a runtime signing key
      description: >
        Evidence already accepted while the key was live stays explainable; new
        submissions signed by it are refused.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [runtime_id]
              properties:
                runtime_id: { type: string }
      responses:
        "200": { description: Revoked }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Operator API key issued by `POST /v1/keys`. Supply as
        `Authorization: Bearer <key>`. The auth middleware extracts the
        `operator_id` from the key row and injects it into the request context.
...
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "missing required field: agent_did"
    Unauthorized:
      description: Missing or invalid Bearer API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "unauthorized"
...